Feature/tron solidity 0.8.0

+++ test/libsolidity/semanticTests/viaYul/dirty_memory_int32.sol
@@ -4,7 +4,7 @@ contract C {
         assembly {
             mstore(m, 0xdeadbeef15dead)
         }
-        int32 x = int32(m[0]);
+        int32 x = int32(uint32(m[0]));
         uint r;
         assembly {
             r := x
@@ -14,5 +14,6 @@ contract C {
 }
 // ====
 // compileViaYul: true
+// compileToEwasm: also
 // ----
 // f() -> true
